Allow passing a byte size window to control how we look for compressed data
Compressed data inside the pack file may be big. Given we need to guess
where a compressed hunk starts and ends, the original implementation
walked the file reader byte by byte, until it would find a valid gzip
buffer. If it found one, then it would proceed to the next hunk.

This approach works, but the bigger the blob objects the more time it
takes to process the whole pack file.

Zlib::inflate can work with an array of bytes that contains multiple
valid chunks of that. When it receives that, it will inflate the first
valid block of data, returning the valid uncompressed content.

However, it cannot process a file pointer. It must receive an array of
bytes.

To mitigate the high processing time but still keep the memory footprint
low, I am introducing a way to control the size of the window we use to
look for data in the packfile. It defaults to 10_000 bytes and can be
set to any value.

With this, we don't need to process byte by byte, but we pay an extra
price when we need to reposition the file pointer back to the end of the
compressed data in the packfile. After finding the compressed data (by
uncompressing it), we need to re-compress the data to figure out its
real size within the packfile.